make doctests work the same on 32/64-bit platforms

+++ scipy/sparse/bsr.py
@@ -62,10 +62,10 @@ class bsr_matrix(_cs_matrix):
 
     >>> from scipy.sparse import *
     >>> from scipy import *
-    >>> bsr_matrix( (3,4), dtype='int32' ).todense()
+    >>> bsr_matrix( (3,4), dtype=int8 ).todense()
     matrix([[0, 0, 0, 0],
             [0, 0, 0, 0],
-            [0, 0, 0, 0]], dtype=int32)
+            [0, 0, 0, 0]], dtype=int8)
 
     >>> row  = array([0,0,1,2,2,2])
     >>> col  = array([0,2,2,0,1,2])
